Updating unit tests and migrating to single file
